Extend tests to test our own generated output
The normal semonitor test outputs a record file which is compared to
the known source.

When using this file as input however, it is also expected that the
outputs match the original. A rec file as input is also not yet tested,
so this extention is two fold.

Contributes to issue #105.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>