Initial commit to add the se2influx.py utility.
The se2influx.py utility will accept semonitor.py input from stdin (allowing
you to use to pipe semonitor.py output into it) and write it to an influxdb instance.

Inverters and optimizers will be written to seperate measurements, with their
unique serial numbers acting as tags, and the rest of the semonitor fields will
be sent verbatim to your influxdb as native fields.

Lastly, se2influx (optionally) allows for writing of ingested semonitor data to
a log file on disk, with built-in log file rotation and compression.  This
allows you to eliminate any uses of "tee" in your piped commands, and permits
you to replay any data that was failed to be sent to your influxdb host (due to
host downtime, etc) at a later date by simplying cat-ing the relevant log
file(s) into the se2influx.py script.

pps.  We also include a "se2influx" script as an example of how you might
want to execute it in conjunction with semonitor.py